+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
 
+## [0.15.1] - TBA
+
+### Added
+
+- New env var: KULA_POSTGRES_PASSWORD
+
 ## [0.15.0] - 2026-04-20
 
 ### Added